+++ src/txt.d.ts
@@ -5,3 +5,8 @@ declare module '*.txt' {
   const content: string
   export default content
 }
+
+declare module '*.md' {
+  const content: string
+  export default content
+}